backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
sqlalchemy==2.0.36
pymysql==1.1.1
alembic==1.13.3
redis==5.0.8
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.10.0
celery==5.4.0
python-multipart==0.0.17
qrcode==7.4.2
python-dotenv==1.0.1
pydantic-settings==2.6.0
httpx==0.27.2
cryptography==43.0.3
python-dateutil==2.9.0.post0
passlib==1.7.4
bcrypt==3.2.2
